Move dss.util.es to dss.index.es
