+++ tests/infra/elasticsearch_test_case.py
@@ -3,7 +3,7 @@ import unittest
 import os
 
 from dss.config import Config
-from dss.util.es import ElasticsearchServer
+from dss.index.es import ElasticsearchServer
 from tests.es import elasticsearch_delete_index
