+++ tests/test_subscriptions.py
@@ -21,7 +21,7 @@ from dss.events.handlers.index import BundleDocument
 from dss.logging import configure_test_logging
 from dss.storage.index import IndexManager
 from dss.util import UrlBuilder
-from dss.util.es import ElasticsearchClient
+from dss.index.es import ElasticsearchClient
 from tests import get_auth_header, get_bundle_fqid
 from tests.infra import DSSAssertMixin, testmode
 from tests.infra.elasticsearch_test_case import ElasticsearchTestCase
